Improve entities selection
To make easier the selection of entities, display their name in addition
of their ID and add search input to easily filter entries inside long
list.

Note : The entities label could now be quite long, so display the
checkbox first. For the same reason, switch to a horizontal menu for
tabs.